add unit tests for new plugins (and rename test file)
